update loading plugin tests
